refactor: split prompt studio services
