+++ tests/fixtures/pyright_baseline.json
@@ -78,19 +78,19 @@
     },
     {
       "count": 3,
-      "path": "easyuse_anima/prompt/regional.py",
+      "path": "easyuse_anima/prompt/regional_builder.py",
       "rule": "reportOptionalIterable",
       "severity": "error"
     },
     {
       "count": 2,
-      "path": "easyuse_anima/prompt/regional.py",
+      "path": "easyuse_anima/prompt/regional_builder.py",
       "rule": "reportOptionalMemberAccess",
       "severity": "error"
     },
     {
       "count": 1,
-      "path": "easyuse_anima/prompt/regional.py",
+      "path": "easyuse_anima/prompt/regional_builder.py",
       "rule": "reportReturnType",
       "severity": "error"
     }
